Add shell extensions to CLI utils class and move into its own module
