+++ .idea/compiler.xml
@@ -26,6 +26,8 @@
       <module name="canonicalizer_test" target="1.8" />
       <module name="client_main" target="1.8" />
       <module name="client_test" target="1.8" />
+      <module name="cliutils_main" target="1.8" />
+      <module name="cliutils_test" target="1.8" />
       <module name="common_main" target="1.8" />
       <module name="common_test" target="1.8" />
       <module name="confidential-identities_main" target="1.8" />
