+++ tools/bootstrapper/build.gradle
@@ -7,6 +7,7 @@ description 'Network bootstrapper'
 
 dependencies {
     compile project(':node-api')
+    compile project(':tools:cliutils')
     compile "info.picocli:picocli:$picocli_version"
     compile "org.slf4j:jul-to-slf4j:$slf4j_version"
     compile "org.apache.logging.log4j:log4j-slf4j-impl:$log4j_version"
