+++ tools/cliutils/src/main/kotlin/net/corda/cliutils/CordaCliWrapper.kt
@@ -1,54 +1,17 @@
-package net.corda.node.utilities
+package net.corda.cliutils
 
-import com.jcabi.manifests.Manifests
-import net.corda.core.internal.exists
-import net.corda.core.internal.isReadable
+import net.corda.core.internal.rootMessage
 import net.corda.core.utilities.contextLogger
+
 import org.apache.logging.log4j.Level
 import picocli.CommandLine
 import picocli.CommandLine.*
-import java.nio.file.Path
 import kotlin.system.exitProcess
 import java.util.*
 
 /**
- * Something heavily used in network services, I am not sure it's of much use in corda, but who knows. Definitely it was the key to making DevOps happy.
- * Add it as
- * `@CommandLine.Mixin
- * lateinit var configParser: ConfigFilePathArgsParser`
- *
- * in your command class and then validate()
- */
-@Command(description = ["Parse configuration file. Checks if given configuration file exists"])
-class ConfigFilePathArgsParser : Validated {
-    @Option(names = ["--config-file", "-f"], required = true, paramLabel = "FILE", description = ["The path to the config file"])
-    lateinit var configFile: Path
-
-    override fun validator(): List<String> {
-        val res = mutableListOf<String>()
-        if(!configFile.exists()) res += "Config file ${configFile.toAbsolutePath().normalize()} does not exist!"
-        if(!configFile.isReadable) res += "Config file ${configFile.toAbsolutePath().normalize()} is not readable"
-        return res
-    }
-}
-
-/**
- * Simple version printing when command is called with --version or -V flag. Assuming that we reuse Corda-Release-Version and Corda-Revision
- * in the manifest file.
- */
-class CordaVersionProvider : IVersionProvider {
-    override fun getVersion(): Array<String> {
-        return if (Manifests.exists("Corda-Release-Version") && Manifests.exists("Corda-Revision")) {
-            arrayOf("Version: ${Manifests.read("Corda-Release-Version")}", "Revision: ${Manifests.read("Corda-Revision")}")
-        } else {
-            arrayOf("No version data is available in the MANIFEST file.")
-        }
-    }
-}
-
-/**
- * Usually when we have errors in some command line flags that are not handled by picocli (e.g. non existing file). Error is thrown
- * and no CommandLine help afterwards. This can be called from run() method.
+ * When we have errors in command line flags that are not handled by picocli (e.g. non existing files), an error is thrown
+ * without any command line help afterwards. This can be called from run() method.
  */
 interface Validated {
     companion object {
@@ -56,6 +19,7 @@ interface Validated {
         const val RED = "\u001B[31m"
         const val RESET = "\u001B[0m"
     }
+
     /**
      * Check that provided command line parameters are valid, e.g. check file existence. Return list of error strings.
      */
@@ -75,6 +39,26 @@ interface Validated {
     }
 }
 
+
+fun CordaCliWrapper.start(vararg args: String) {
+    val cmd = CommandLine(this)
+    cmd.commandSpec.name(alias)
+    cmd.commandSpec.usageMessage().description(description)
+    try {
+        cmd.parseWithHandlers(RunLast().useOut(System.out).useAnsi(Help.Ansi.AUTO),
+                DefaultExceptionHandler<List<Any>>().useErr(System.err).useAnsi(Help.Ansi.AUTO),
+                *args)
+    } catch (e: ExecutionException) {
+        val throwable = e.cause ?: e
+        if (this.verbose) {
+            throwable.printStackTrace()
+        } else {
+            System.err.println("*ERROR*: ${throwable.rootMessage ?: "Use --verbose for more details"}")
+        }
+        exitProcess(1)
+    }
+}
+
 /**
  * Simple base class for handling help, version, verbose and logging-level commands.
  * As versionProvider information from the MANIFEST file is used. It can be overwritten by custom version providers (see: Node)
@@ -83,13 +67,14 @@ interface Validated {
 @Command(mixinStandardHelpOptions = true,
         versionProvider = CordaVersionProvider::class,
         sortOptions = false,
+        showDefaultValues = true,
         synopsisHeading = "%n@|bold,underline Usage|@:%n%n",
         descriptionHeading = "%n@|bold,underline Description|@:%n%n",
         parameterListHeading = "%n@|bold,underline Parameters|@:%n%n",
         optionListHeading = "%n@|bold,underline Options|@:%n%n",
         commandListHeading = "%n@|bold,underline Commands|@:%n%n")
-abstract class ArgsParser {
-    @Option(names = [ "-v", "--verbose" ], description = ["If set, prints logging to the console as well as to a file."])
+abstract class CordaCliWrapper(val alias: String, val description: String) : Runnable {
+    @Option(names = ["-v", "--verbose"], description = ["If set, prints logging to the console as well as to a file."])
     var verbose: Boolean = false
 
     @Option(names = ["--logging-level"],
@@ -98,14 +83,26 @@ abstract class ArgsParser {
             converter = [LoggingLevelConverter::class])
     var loggingLevel: Level = Level.INFO
 
+    @Mixin
+    private lateinit var installShellExtensionsParser: InstallShellExtensionsParser
+
     // This needs to be called before loggers (See: NodeStartup.kt:51 logger called by lazy, initLogging happens before).
     // Node's logging is more rich. In corda configurations two properties, defaultLoggingLevel and consoleLogLevel, are usually used.
-    protected open fun initLogging() {
+    private fun initLogging() {
         val loggingLevel = loggingLevel.name().toLowerCase(Locale.ENGLISH)
         System.setProperty("defaultLogLevel", loggingLevel) // These properties are referenced from the XML config file.
         if (verbose) {
             System.setProperty("consoleLogLevel", loggingLevel)
         }
+    }
+
+    // Override this function with the actual method to be run once all the arguments have been parsed
+    abstract fun invoke()
+
+    override fun run() {
+        installShellExtensionsParser.installOrUpdateShellExtensions(alias)
+        initLogging()
+        invoke()
     }
 }
 
@@ -114,6 +111,7 @@ abstract class ArgsParser {
  */
 class LoggingLevelConverter : ITypeConverter<Level> {
     override fun convert(value: String?): Level {
-        return value?.let { Level.getLevel(it) } ?: throw TypeConversionException("Unknown option for --logging-level: $value")
+        return value?.let { Level.getLevel(it) }
+                ?: throw TypeConversionException("Unknown option for --logging-level: $value")
     }
 }
